KS-1-4 show values 2

+++ src/main/kotlin/lesson_1/KS-1-4.kt
@@ -3,9 +3,9 @@ package org.example.lesson_1
 fun main ()
 {
     val length: String = "40_868_600_000"
-    val age: Short = 27
-    val partOfDay: Double = 0.075
-    val seconds: Int = 6480
+    val age: Byte = 27
+    val partOfDay: Float = 0.075f
+    val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
